Integrate raise hand sound into call event renderer.
