+++ src/room/CallEventAudioRenderer.tsx
@@ -5,14 +5,17 @@ SPDX-License-Identifier: AGPL-3.0-only
 Please see LICENSE in the repository root for full details.
 */
 
-import { ReactNode, useEffect } from "react";
+import { ReactNode, useDeferredValue, useEffect, useMemo } from "react";
 import { debounce, filter, interval, throttle } from "rxjs";
 import { CallViewModel } from "../state/CallViewModel";
 import joinCallSoundMp3 from "../sound/join_call.mp3";
 import joinCallSoundOgg from "../sound/join_call.ogg";
 import leftCallSoundMp3 from "../sound/left_call.mp3";
 import leftCallSoundOgg from "../sound/left_call.ogg";
+import handSoundOgg from "../sound/raise_hand.ogg?url";
+import handSoundMp3 from "../sound/raise_hand.mp3?url";
 import { prefetchSounds, useAudioContext } from "../useAudioContext";
+import { useReactions } from "../useReactions";
 
 // Do not play any sounds if the participant count has exceeded this
 // number.
@@ -29,6 +32,10 @@ const Sounds = prefetchSounds({
     mp3: leftCallSoundMp3,
     ogg: leftCallSoundOgg,
   },
+  raiseHand: {
+    mp3: handSoundMp3,
+    ogg: handSoundOgg,
+  },
 });
 
 export function CallEventAudioRenderer({
@@ -40,6 +47,19 @@ export function CallEventAudioRenderer({
     sounds: Sounds,
     latencyHint: "interactive",
   });
+
+  const { raisedHands } = useReactions();
+  const raisedHandCount = useMemo(
+    () => Object.keys(raisedHands).length,
+    [raisedHands],
+  );
+  const previousRaisedHandCount = useDeferredValue(raisedHandCount);
+
+  useEffect(() => {
+    if (audioEngineCtx && previousRaisedHandCount < raisedHandCount) {
+      audioEngineCtx.playSound("raiseHand");
+    }
+  }, [audioEngineCtx, previousRaisedHandCount, raisedHandCount]);
 
   useEffect(() => {
     if (!audioEngineCtx) {
